♻️ Make a function to handle board square selection
