+++ src/states/settings.ts
@@ -1,3 +1,5 @@
+import Draw from '~/game/draw';
+
 /** Whether the app makes a sound. */
 let isMuted: boolean = true;
 
@@ -19,3 +21,12 @@ const toggleShowOppositePieces = () => {
 export const showOppositePiecesValue = () => showOppositePieces;
 
 export const useShowOppositePieces = () => ({ showOppositePieces, toggleShowOppositePieces });
+
+/** `Draw` instance. */
+let draw: Draw;
+
+export const drawValue = () => draw;
+
+export const setDraw = (state: Draw) => {
+  draw = state;
+};
